+++ service/grunt/clean.js
@@ -1,0 +1,5 @@
+module.exports = {
+  coverage: {
+    src: ['test/coverage/']
+  }
+};