+++ service/grunt/env.js
@@ -1,0 +1,5 @@
+module.exports = {
+  coverage: {
+    APP_DIR_FOR_CODE_COVERAGE: '../test/coverage/instrument/lib/'
+  }
+};