+++ service/grunt/instrument.js
@@ -1,0 +1,7 @@
+module.exports = {
+  files: 'lib/**/*.js',
+  options: {
+    lazy: true,
+    basePath: 'test/coverage/instrument/'
+  }
+};