+++ service/grunt/storeCoverage.js
@@ -1,0 +1,5 @@
+module.exports = {
+  options: {
+    dir: 'test/coverage/reports'
+  }
+};